# optimize ft_volumelookup computing time