--- a/ft_volumelookup.m
+++ b/ft_volumelookup.m
@@ -1,31 +1,31 @@
 function [output] = ft_volumelookup(cfg, volume)
 
-% FT_VOLUMELOOKUP can be used in to combine an anatomical or functional atlas with
-% the source reconstruction results. You can use it for forward and reverse lookup.
+% FT_VOLUMELOOKUP can be used in to combine an anatomical or functional
+% atlas with the source reconstruction results. You can use it for forward
+% and reverse lookup.
 %
-% Given the ROI as anatomical or functional label, it looks up the locations and
-% creates a mask (as a binary volume) based on the label. Given the ROI as point in
-% the brain, it creates a sphere or box around that point. In these two case the
-% function is to be used as:
+% Given the region of interest (ROI) as anatomical or functional label, it
+% looks up the locations and creates a mask (as a binary volume) based on
+% the label. Given the ROI as point in the brain, it creates a sphere or
+% box around that point. In these two case the function is to be used as:
 %   mask = ft_volumelookup(cfg, volume)
 %
-% Given a binary volume that indicates a region of interest or a point of
-% interest, it looks up the corresponding anatomical or functional labels
-% from the atlas. In this case the function is to be used as:
+% Given a binary volume that indicates a ROI or a point of interest (POI),
+% it looks up the corresponding anatomical or functional labels from the
+% atlas. In this case the function is to be used as:
 %    labels = ft_volumelookup(cfg, volume)
 %
 % In both cases the input volume can be:
-%   mri    is the output of FT_READ_MRI
-%   source is the output of FT_SOURCEANALYSIS
+%   mri    is the output of FT_READ_MRI source is the output of FT_SOURCEANALYSIS
 %   stat   is the output of FT_SOURCESTATISTICS
 %
 % The configuration options for a mask according to an atlas:
 %   cfg.inputcoord          = 'mni' or 'tal', coordinate system of the mri/source/stat
 %   cfg.atlas               = string, filename of atlas to use, see FT_READ_ATLAS
-%   cfg.roi                 = string or cell-array of strings, region(s) of interest from anatomical atlas
+%   cfg.roi                 = string or cell-array of strings, ROI from anatomical atlas
 %
-% The configuration options for a spherical/box mask around a point of interest:
-%   cfg.roi                 = Nx3 vector, coordinates of the points of interest
+% The configuration options for a spherical/box mask around a POI:
+%   cfg.roi                 = Nx3 vector, coordinates of the POI
 %   cfg.sphere              = radius of each sphere in cm/mm dep on unit of input
 %   cfg.box                 = Nx3 vector, size of each box in cm/mm dep on unit of input
 %   cfg.round2nearestvoxel  = 'yes' or 'no' (default = 'no'), voxel closest to point of interest is calculated
@@ -37,16 +37,17 @@ function [output] = ft_volumelookup(cfg, volume)
 %   cfg.maskparameter       = string, field in volume to be looked up, data in field should be logical
 %   cfg.maxqueryrange       = number, should be odd (default = 1)
 %
-% The configuration options for labels around a point of interest:
+% The configuration options for labels around POI:
 %   cfg.output              = 'label'
-%   cfg.roi                 = Nx3 vector, coordinates of the points of interest
+%   cfg.roi                 = Nx3 vector, coordinates of the POI
 %   cfg.inputcoord          = 'mni' or 'tal', coordinate system of the mri/source/stat
 %   cfg.atlas               = string, filename of atlas to use, see FT_READ_ATLAS
 %   cfg.maxqueryrange       = number, should be 1, 3, 5 (default = 1)
-%   cfg.querymethod         = 'sphere' searches voxels around the roi in a sphere (default)
-%                           = 'cube' searches voxels around the roi in a sphere
-%   cfg.round2nearestvoxel  = 'yes' or 'no', voxel closest to point of interest is calculated (default = 'yes')
-%   cfg.multioutput         = 'yes' or 'no', one output per point of interest (default = 'no')
+%   cfg.querymethod         = 'sphere' searches voxels around the ROI in a sphere (default)
+%                           = 'cube' searches voxels around the ROI in a cube
+%   cfg.round2nearestvoxel  = 'yes' or 'no', voxel closest to POI is calculated (default = 'yes')
+%   cfg.multioutput         = 'no' always outputs one label; if several POI are provided, they are considered together as describing a ROI (default)
+%                             'yes' outputs one label per POI (e.g., choose to get labels for different electrodes)
 %
 % The label output has a field "names", a field "count" and a field "usedqueryrange".
 % To get a list of areas of the given mask you can do for instance:
@@ -349,7 +350,7 @@ elseif mask2label || roi2label
     if istrue(cfg.round2nearestvoxel)
       % determine location of each anatomical voxel in head coordinates
       xyz = [volume.pos ones(size(volume.pos,1),1)]'; % note that this is 4xN
-      nSel = size(cfg.roi,1);
+      nSel = size(cfg.roi, 1);
       for i=1:nSel
         cfg.roi(i,:) = poi2voi(cfg.roi(i,:), xyz);
       end
@@ -360,9 +361,9 @@ elseif mask2label || roi2label
     end
   end
   if strcmp(cfg.multioutput, 'yes')
-      labels = repmat(labels, nSel, 1);
+      labels = repmat(labels, length(sel), 1);
   end
-  for iVox = 1:nSel
+  for iVox = 1:length(sel)
       label = {};
       for qr = 1:2:cfg.maxqueryrange
           if isempty(label)