--- a/test/test_ft_volumelookup.m
+++ b/test/test_ft_volumelookup.m
@@ -50,3 +50,14 @@ cfg.inputcoord = 'mni';
 cfg.roi = 'Calcarine_R'; % right V1
 mask6 = ft_volumelookup(cfg, mri);
 assert(isequal(sum(mask6(:)),1861));
+
+atlas_MNI = ft_read_atlas(atlasfilename);
+atlas_MNI.coordsys = 'mni';
+cfg = [];
+cfg.roi = [52 -9 -45; 73 -37 -8];
+cfg.atlas = atlasfilename;
+cfg.inputcoord = 'mni';
+cfg.output = 'label';
+cfg.maxqueryrange = 29;
+cfg.multioutput = 'yes';
+labels_MNI = ft_volumelookup(cfg, atlas_MNI);