Hi there,

Currently working on a 250-channel ECoG grid case, I used ft_volumelookup to get each channel's anatomical label, and it took about 1h37min (~23.2 seconds per channel), which was surprisingly long to me. After a two-step optimization, it now takes slightly more than 7 minutes =)

First, the ismember call on line 354 was replaced by a faster implementation (ismember is actually [known to be slow](https://fr.mathworks.com/matlabcentral/answers/7434-find-vector-in-matrix)); running this line was then taking 0.57 seconds instead of ~19 seconds;

However, running this new version of ft_volumelookup was then taking ~6 seconds (thus 25 minutes for my 250 channels), which meant that this line wasn't the only time-consuming step within ft_volumelookup, thus that there was still room for optimization. Indeed, some channel-/point-of-interest-independent steps were making most of this computing time (e.g., ft_checkdata on line 152 was taking 2.6 seconds; a big matrix transposition of line 351 was taking 1.7 seconds): to get labels for my 250 channels, I was thus unnecessarily repeating more than 4.3 seconds steps, for a total of about 18 unnecessary minutes.

Therefore, second optimization step consisted of enabling a multioutput option, considering rows of cfg.roi as independent points of interest, and giving one output by channel; that way, POI-independent steps are done once for all POIs, and this optimized version of ft_volumelookup now takes about 7 minutes to run.

I implemented this option so that previous calls to ft_volumelookup wouldn't be affected by it (default cfg.multioutput is 'no'). In a way, cfg.multioutput is a switch to either see cfg.roi as points defining a region of interest (when equal 'no' => "integrated", single output), or as a list of individual points (when equal 'yes' => same output structure but with as many levels as the number of points provided).

Best,
Ludovic

PS: absolute time values I gave here to illustrate this pull request obviously depend on the resolution of the atlas/volume used, but a general claim would be that this two-step optimization improves computing time of ft_volumelookup by (at least) one order of magnitude.