@robertoostenveld 2018-05-08
Speed improvements are welcome! However, with your suggested change test_ft_volumelookup (in the test directory) starts failing 

```
Undefined function or variable 'nSel'.
Error in ft_volumelookup (line 365)
  for iVox = 1:nSel
Error in test_ft_volumelookup (line 43)
mask5     = ft_volumelookup(cfg, mri); 
```

Can you add your usage pattern to the test function and ensure that it does not fail on the older use cases?

Furthermore, it is unclear what the new multioutput option does due to lack of documentation. What would the difference be in the output when it is yes or no?

I think that the overall documentation of the function is poor. It seems that the options can be described as a 2x2 table (2 outputs, each with 2 options for the inputs). Better would be to change it such that the goals are more clear. So rather than "The configuration options for ... " the help  should say "The configuration options to obtain xxx from yyy are ...". Can you update the help? And can you fix querymethod cube/shpere in the help?