@ludovicbellier 2018-05-08
The 'nSel' error is now corrected, and test_ft_volumelookup.m is updated with my new usage pattern.

About the documentation, here's the rationale of this multioutput option:
- if you provide several points of interest to the original version of ft_volumelookup.m, these POI are considered as describing a region of interest (hence the field name, cfg.roi) you get only one label structure;
- if you provide several POI to this new version of ft_volumelookup.m, you can keep its previous behavior by choosing cfg.multioutput = 'no', or you can tell the function to consider each POI as independent by choosing cfg.multioutput = 'yes' and then to give you one label per POI (that's especially of interest in the case you want labels for many electrodes, as this is the main factor for speed improvement).

As the two usages are very close, I updated the documentation trying to be the clearest possible, but without repeating a pretty big paragraph which would make the documentation heavier.